Clean up JSON cast related methods in TypeJsonUtils
JSON cast are rewritten for better efficiency, these utility methods
are in JsonUtil.

In this commit:
- Remove unused methods in TypeJsonUtils
- Refactor canCastFromJson into JsonUtil
