+++ presto-main/src/main/java/com/facebook/presto/operator/scalar/JsonToMapCast.java
@@ -41,11 +41,11 @@ import static com.facebook.presto.metadata.Signature.comparableTypeParameter;
 import static com.facebook.presto.metadata.Signature.typeVariable;
 import static com.facebook.presto.spi.StandardErrorCode.INVALID_CAST_ARGUMENT;
 import static com.facebook.presto.spi.type.TypeSignature.parseTypeSignature;
-import static com.facebook.presto.type.TypeJsonUtils.canCastFromJson;
 import static com.facebook.presto.util.Failures.checkCondition;
 import static com.facebook.presto.util.JsonUtil.BlockBuilderAppender.createBlockBuilderAppender;
 import static com.facebook.presto.util.JsonUtil.HashTable;
 import static com.facebook.presto.util.JsonUtil.JSON_FACTORY;
+import static com.facebook.presto.util.JsonUtil.canCastFromJson;
 import static com.facebook.presto.util.JsonUtil.createJsonParser;
 import static com.facebook.presto.util.JsonUtil.truncateIfNecessaryForErrorMessage;
 import static com.facebook.presto.util.Reflection.methodHandle;
